Time: 1 ms (81.54%) | Memory: 41.2 MB (23.49%) - LeetSync

+++ 7-reverse-integer/reverse-integer.java
@@ -24,23 +24,21 @@ class Solution {
         //  }
         // return 0;
 
-        String a = Math.abs(x) + ""; // use long to avoid overflow on abs
+        String a = Math.abs(x) + ""; 
         long rev = 0;
 
-        long temp = Math.abs((long) x); // work with positive number
+        long temp = Math.abs(x); // work with positive number
         for (int i = a.length() - 1; i >= 0; i--) {
-            int ld = (int) (temp % 10); // extract last digit
-            long num = (long) Math.pow(10, i); // place value
+            int ld = (int) (temp % 10);
+            long num = (long) Math.pow(10, i); 
             rev += num * ld;
             temp /= 10;
         }
 
-        // apply original sign
         if (x < 0) {
             rev = -rev;
         }
 
-        // check if within int range
         if (rev >= Integer.MIN_VALUE && rev <= Integer.MAX_VALUE) {
             return (int) rev;
         }
